requirements.txt
# ⚠️ IMPORTANT:
# This project is compatible with:
# ✅ Python 3.9, 3.10, or 3.11
# ❌ NOT compatible with Python 3.12+ (pyannote will break)

# -----------------------------
# Core ML & Audio Processing
# -----------------------------

torch==2.1.2
torchaudio==2.1.2

numpy==1.26.4

pyannote.audio==3.1.1
pyannote.core<6.0.0
pyannote.metrics<4.0.0

faster-whisper==1.0.3

huggingface-hub<0.20

soundfile
ffmpeg-python


# -----------------------------
# Backend / API
# -----------------------------

flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0


# -----------------------------
# Utilities
# -----------------------------

tqdm
